+++ src/main/java/com/Twitter/org/services/Replies/RepliesServiceImpl.java
@@ -3,9 +3,12 @@ package com.Twitter.org.services.Replies;
 import com.Twitter.org.Models.Response;
 import com.Twitter.org.Models.Tweets.Replies;
 import com.Twitter.org.Models.Tweets.Tweets;
+import com.Twitter.org.Models.dto.TweetsDto.TweetsCreateDto;
 import com.Twitter.org.Repository.RepliesRepository;
 import com.Twitter.org.Repository.TweetsRepository;
-import com.Twitter.org.mappers.Impl.TweetsMapper;
+import com.Twitter.org.mappers.Impl.TweetsMapper.TweetsDetailsDtoMapper;
+import com.Twitter.org.mappers.Mapper;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,13 +21,15 @@ import java.util.Optional;
 public class RepliesServiceImpl implements RepliesService {
 
     private final TweetsRepository tweetsRepository;
-    private final TweetsMapper TweetsMapper;
     private final RepliesRepository repliesRepository;
+    private final Mapper<Tweets, TweetsCreateDto> tweetsCreateDtoMapper;
+    private final TweetsDetailsDtoMapper tweetsDetailsDtoMapper;
 
-    public RepliesServiceImpl(TweetsRepository tweetsRepository, com.Twitter.org.mappers.Impl.TweetsMapper tweetsMapper, RepliesRepository repliesRepository) {
+    public RepliesServiceImpl(TweetsRepository tweetsRepository, RepliesRepository repliesRepository, Mapper<Tweets, TweetsCreateDto> tweetsCreateDtoMapper, TweetsDetailsDtoMapper tweetsDetailsDtoMapper) {
         this.tweetsRepository = tweetsRepository;
-        TweetsMapper = tweetsMapper;
         this.repliesRepository = repliesRepository;
+        this.tweetsCreateDtoMapper = tweetsCreateDtoMapper;
+        this.tweetsDetailsDtoMapper = tweetsDetailsDtoMapper;
     }
 
     Response validParent(int parentId) {
@@ -40,69 +45,87 @@ public class RepliesServiceImpl implements RepliesService {
         return response;
     }
 
+    @Transactional
     @Override
-    public Response comment(Tweets comment) {
-        // 1. Validate the request
-        int parentTweetId = comment.getParentId();
-        Response response = validParent(parentTweetId);
+    public Response comment(@NotNull TweetsCreateDto comment) {
+        Response response = validParent(comment.getParentId());
         if (!response.isSuccess()) {
             return response;
         }
 
-        // 2. Create the entity
         Tweets parentTweet = (Tweets) response.getData();
         parentTweet.setRepliesNumber(parentTweet.getRepliesNumber() + 1);
         tweetsRepository.save(parentTweet);
 
-        comment.setOriginalPost(parentTweetId); // Set the originalPostId of the comment
-        comment.setComment(true); // Set the comment to true
+        Tweets newComment = tweetsCreateDtoMapper.mapFrom(comment);
+        newComment.setOriginalPost(comment.getParentId());
+        newComment.setComment(true);
+        newComment.setId(null);
+        tweetsRepository.save(newComment);
 
-        // 3. Save the entity
-        comment.setId(null); // Ignore the id if it is supplied, To avoid conflicts
-        tweetsRepository.save(comment);
-
-        // 4. Create the Replies object
-        Replies reply = new Replies(parentTweet, comment);
+        Replies reply = new Replies(parentTweet, newComment);
         repliesRepository.save(reply);
 
-        // 6. Return a response
         Map<String, Object> data = new HashMap<>();
+        data.put("parentTweet", tweetsDetailsDtoMapper.mapTo(parentTweet));
+        data.put("comment", tweetsDetailsDtoMapper.mapTo(newComment));
 
-        // To fix cyclic reference, we will map the parent tweet to a DTO object
-        data.put("parentTweet", TweetsMapper.mapTo(parentTweet));
-        data.put("comment", TweetsMapper.mapTo(comment));
         return new Response(true, "Comment created successfully!", data);
     }
 
+    @Transactional
     @Override
     public Response deleteComment(int commentTweetId) {
         Response response = new Response();
-        Optional<Tweets> comment = tweetsRepository.findById(commentTweetId);
-        if (comment.isEmpty() || !comment.get().getComment()) {
+
+        // Find the comment tweet
+        Optional<Tweets> commentOptional = tweetsRepository.findById(commentTweetId);
+
+        // If the comment does not exist
+        if (commentOptional.isEmpty()) {
             response.setSuccess(false);
-            response.setMessage("The comment does not exist!");
+            response.setMessage("The comment does not exist or is not a valid comment!");
             return response;
         }
 
-        Tweets parentTweet = comment.get();
-        int parentId = parentTweet.getParentId();
-        Optional<Tweets> parentTweetOptional = tweetsRepository.findById(parentId);
-        String message = "";
-        if (parentTweetOptional.isEmpty()) {
-            response.setMessage("Note, The parent tweet does not exist!\n");
-        } else {
-            parentTweet = parentTweetOptional.get();
-            parentTweet.setRepliesNumber(parentTweet.getRepliesNumber() - 1);
-            tweetsRepository.save(parentTweet);
-            response.setData(TweetsMapper.mapTo(parentTweet));
+        Tweets commentTweet = commentOptional.get();
+        Integer parentId = commentTweet.getParentId();
+
+        // If isComment is false, state that the tweet is not a comment
+        if (!commentTweet.getComment()) {
+            response.setSuccess(false);
+            response.setMessage("The tweet is not a comment!, Use different endpoint to delete the tweet");
+            return response;
         }
 
-        tweetsRepository.deleteById(commentTweetId); // Deleting the comment Tweet will delete the reply object as well
+        // If parentId is null, set a message and delete the comment tweet
+        if (parentId == null) {
+            response.setMessage("Note, The parent tweet id is null!\n");
+            tweetsRepository.deleteById(commentTweetId);
+            response.setSuccess(true);
+            response.setMessage(response.getMessage() + "Comment deleted successfully!");
+            return response;
+        }
+
+        // Find the parent tweet
+        Optional<Tweets> parentTweetOptional = tweetsRepository.findById(parentId);
+        if (parentTweetOptional.isPresent()) {
+            Tweets parentTweet = parentTweetOptional.get();
+            parentTweet.setRepliesNumber(parentTweet.getRepliesNumber() - 1);
+            tweetsRepository.save(parentTweet);
+            response.setData(tweetsDetailsDtoMapper.mapTo(parentTweet));
+        } else {
+            response.setMessage("Note, The parent tweet does not exist!\n");
+        }
+
+        // Delete the comment tweet
+        tweetsRepository.deleteById(commentTweetId);
+
         response.setSuccess(true);
-        message += "Comment deleted successfully!";
-        response.setMessage(message);
+        response.setMessage(response.getMessage() + "Comment deleted successfully!");
         return response;
     }
+
 
     @Override
     public int countReplies(int tweetId) {
